+++ frontends/web/ts/index.ts
@@ -1,3 +1,5 @@
+export * from "./element";
+export * from "./embed";
 export * from "./gb";
 export * from "./palettes";
 export * from "./playlist";